requirements.txt
streamlit==1.38.0
pandas==2.2.2
supabase==2.5.1
pdfplumber==0.11.4
python-docx==1.1.2
spacy==3.7.5
sentence-transformers==2.2.2
numpy==1.24.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
